First commit
Made a brick breaker game using C# in the .NET framework. I did this by using observable collections for the bricks, events for when the ball hit either a paddle or a brick, and timers to control the movement of the ball. This was on done in the context of a WPA application.